requirements.txt
flask==2.0.1
flask-restx==0.5.1
flask-sqlalchemy==2.5.1
flask-cors==3.0.10
python-dotenv==0.19.0
openai==0.27.8
mapbox==0.18.1
requests==2.31.0
beautifulsoup4==4.12.2
SQLAlchemy==1.4.23
PyJWT==2.1.0
python-dateutil==2.8.2
pydantic==1.10.13
Werkzeug==2.0.1
gunicorn==20.1.0 